+++ materialList/src/main/java/com/dexafree/materialList/listeners/RecyclerItemClickListener.java
@@ -1,4 +1,4 @@
-package com.dexafree.materialList.controller;
+package com.dexafree.materialList.listeners;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
