+++ materialList/src/main/java/com/dexafree/materialList/listeners/SwipeDismissRecyclerViewTouchListener.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.dexafree.materialList.controller;
+package com.dexafree.materialList.listeners;
 
 import android.graphics.Rect;
 import android.os.SystemClock;
